+++ app/lib/utils/index.ts
@@ -1,0 +1,29 @@
+/**
+ * Centralized utility exports
+ *
+ * This file re-exports all utilities from their domain-specific modules.
+ */
+
+// Array utilities
+export * from './array'
+
+// Formatting utilities
+export * from './formatting'
+
+// Date utilities
+export * from './dates'
+
+// String utilities
+export * from './strings'
+
+// DOM utilities
+export * from './dom'
+
+// Type guards
+export * from './guards'
+
+// Miscellaneous utilities
+export * from './misc'
+
+// Re-export viewport helpers from constants
+export { isMobile, isTablet, isDesktop, BREAKPOINTS } from '../constants'
